# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DokMart - Digital document marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Document {
  id: number;
  title: string;
  description: string;
  subject: string;
  level: string;
  /** @nullable */
  semester?: string | null;
  /** @nullable */
  docType?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  price: number;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  isFeatured: boolean;
  /** @nullable */
  pageCount?: number | null;
  downloadCount: number;
  createdAt: string;
}

export interface ListDocumentsResponse {
  documents: Document[];
  total: number;
  page: number;
  totalPages: number;
}

export interface CreateDocumentBody {
  title: string;
  description: string;
  subject: string;
  level: string;
  /** @nullable */
  semester?: string | null;
  /** @nullable */
  docType?: string | null;
  /** @nullable */
  categoryId?: number | null;
  price: number;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  isFeatured?: boolean;
  /** @nullable */
  pageCount?: number | null;
}

export interface UpdateDocumentBody {
  title?: string;
  description?: string;
  subject?: string;
  level?: string;
  /** @nullable */
  semester?: string | null;
  /** @nullable */
  docType?: string | null;
  /** @nullable */
  categoryId?: number | null;
  price?: number;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  isFeatured?: boolean;
  /** @nullable */
  pageCount?: number | null;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  documentCount: number;
}

export interface CreateCategoryBody {
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface Level {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder: number;
  documentCount: number;
}

export type ListLevelsResponse = Level[];

export interface CreateLevelBody {
  name: string;
  slug: string;
  /** @nullable */
  group?: string | null;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
}

export interface DeleteLevelParams {
  id: number;
}

export interface OrderItem {
  id: number;
  documentId: number;
  documentTitle: string;
  documentSubject: string;
  documentLevel: string;
  price: number;
  /** @nullable */
  fileUrl?: string | null;
}

export interface Order {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  totalAmount: number;
  status: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface OrderWithItems {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  totalAmount: number;
  status: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  updatedAt: string;
  items: OrderItem[];
}

export type CreateOrderBodyItemsItem = {
  documentId: number;
};

export interface CreateOrderBody {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  items: CreateOrderBodyItemsItem[];
}

export interface UploadPaymentProofBody {
  customerEmail: string;
  paymentMethod: string;
  proofImageData: string;
  /** @nullable */
  notes?: string | null;
}

export interface AdminOrderDetail {
  id: number;
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  totalAmount: number;
  status: string;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  updatedAt: string;
  items: OrderItem[];
  /** @nullable */
  proofImageData?: string | null;
  /** @nullable */
  proofNotes?: string | null;
  /** @nullable */
  proofUploadedAt?: string | null;
}

export interface ListAdminOrdersResponse {
  orders: Order[];
  total: number;
  page: number;
  totalPages: number;
}

export interface ReviewOrderBody {
  status: string;
  /** @nullable */
  adminNote?: string | null;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export interface DocumentFile {
  id: number;
  documentId: number;
  objectPath: string;
  fileName: string;
  /** @nullable */
  fileSize?: number | null;
  sortOrder: number;
  createdAt: string;
}

export interface AddDocumentFileBody {
  objectPath: string;
  fileName: string;
  fileSize?: number;
  sortOrder?: number;
}

export interface GetDocumentFilesParams {
  id: number;
}

export interface DeleteDocumentFileParams {
  id: number;
  fileId: number;
}

export interface SellerApplicationBody {
  name: string;
  email: string;
  phone: string;
  productType: string;
  description: string;
  /** @nullable */
  portfolioUrl?: string | null;
  password: string;
}

export interface SellerApplication {
  id: number;
  name: string;
  email: string;
  phone: string;
  productType: string;
  description: string;
  /** @nullable */
  portfolioUrl?: string | null;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}

export interface ListSellerApplicationsResponse {
  applications: SellerApplication[];
  total: number;
}

export interface ReviewSellerApplicationBody {
  status: string;
  /** @nullable */
  adminNote?: string | null;
}

export interface AdminStats {
  totalDocuments: number;
  totalOrders: number;
  pendingPayments: number;
  approvedOrders: number;
  totalRevenueCfa: number;
  recentOrders: Order[];
  topDocuments: Document[];
}

export type ListSellerApplicationsParams = {
  status?: string;
};

export type ListDocumentsParams = {
  category?: string;
  level?: string;
  search?: string;
  page?: number;
  limit?: number;
};

export type GetOrderParams = {
  email: string;
};

export type ListAdminOrdersParams = {
  status?: string;
  page?: number;
  limit?: number;
};