# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DokMart - Digital document marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/{id}/files:
    get:
      operationId: getDocumentFiles
      tags: [documents]
      summary: List files attached to a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of document files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentFile"
    post:
      operationId: addDocumentFile
      tags: [documents]
      summary: Attach a file to a document (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDocumentFileBody"
      responses:
        "201":
          description: Created file entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentFile"
  /documents/{id}/files/{fileId}:
    delete:
      operationId: deleteDocumentFile
      tags: [documents]
      summary: Remove a file from a document (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create a new category (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCategoryBody"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
...
  /categories/{id}:
    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete a category (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders for admin
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of all orders with customer info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListAdminOrdersResponse"
  /admin/orders/{id}:
    get:
      operationId: getAdminOrder
      tags: [admin]
      summary: Get order details for admin (with payment proof)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order with payment proof
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: reviewOrder
      tags: [admin]
      summary: Approve or reject an order (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        documentCount:
          type: integer
      required:
        - id
        - name
        - slug
        - documentCount
    CreateCategoryBody:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
      required:
        - name
        - slug
...
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentTitle:
          type: string
        documentSubject:
          type: string
        documentLevel:
          type: string
        price:
          type: integer
        fileUrl:
          type: ["string", "null"]
      required:
        - id
        - documentId
        - documentTitle
        - documentSubject
        - documentLevel
        - price
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        totalAmount:
          type: integer
        status:
          type: string
        paymentMethod:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - totalAmount
        - status
        - createdAt
        - updatedAt
...
    AdminOrderDetail:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
        totalAmount:
          type: integer
        status:
          type: string
        paymentMethod:
          type: ["string", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        proofImageData:
          type: ["string", "null"]
        proofNotes:
          type: ["string", "null"]
        proofUploadedAt:
          type: ["string", "null"]
      required:
        - id
        - customerName
        - customerEmail
        - customerPhone
        - totalAmount
        - status
        - createdAt
        - updatedAt
        - items
    ListAdminOrdersResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - orders
        - total
        - page
        - totalPages
    ReviewOrderBody:
      type: object
      properties:
        status:
          type: string
        adminNote:
          type: ["string", "null"]
      required:
        - status
...
    DocumentFile:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        objectPath:
          type: string
        fileName:
          type: string
        fileSize:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - documentId
        - objectPath
        - fileName
        - sortOrder
        - createdAt
    AddDocumentFileBody:
      type: object
      properties:
        objectPath:
          type: string
        fileName:
          type: string
        fileSize:
          type: integer
        sortOrder:
          type: integer
          default: 0
      required:
        - objectPath
        - fileName
...